requirements.txt
# --- Framework y Servidor ---
django
djangorestframework
djangorestframework-simplejwt

# --- Tareas en Segundo Plano ---
celery[redis]

# --- Base de Datos ---
psycopg2-binary

# --- Lógica de LangChain y Gemini ---
langchain
langchain-google-genai
langchain_community

# --- Gestión de Variables de Entorno ---
python-dotenv



beautifulsoup4
langchain-tavily
langgraph
tavily-python
langchain-openai

feedparser